-Set warning level 4 -Fixed a bunch of compiler warnings
git-svn-id: https://fuji.googlecode.com/svn/trunk@721 7e75b669-355e-e21c-2746-f45317fb21cf

+++ Fuji/Private/Include/MFCallstack.h
@@ -68,7 +68,7 @@ public:
 #endif
 
 #if defined(_MFCALLSTACK_PROFILING)
-	#define MFCALLSTACK_PROFILE(name, mode) static MFCallstack_Function MFCallstack_FunctionDef = { 0LL, name, 0, MFCallstack_GetNextColour(), (mode), false }; MFCall _call(name, &MFCallstack_FunctionDef)
+	#define MFCALLSTACK_PROFILE(name, mode) static MFCallstack_Function MFCallstack_FunctionDef = { 0LL, name, 0, (uint16)MFCallstack_GetNextColour(), (mode), false }; MFCall _call(name, &MFCallstack_FunctionDef)
 #else
 	#define MFCALLSTACK_PROFILE(name, mode) MFCall _call(name, NULL)
 #endif
